fix(lib): rename CommonGenerator and RandIntGenerator files to be in kebabcase
