+++ src/lib/mock-observations/helpers/common-generator.ts
